pyproject.toml
[tool.poetry]
name = "aitk"
version = "0.1.0"
description = "AI Toolkit"
authors = ["Jeff Lee <lgf2111@gmail.com>"]
readme = "README.md"
packages = [{include = "aitk"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.111.1"
pillow = "^10.4.0"
pydantic-settings = "^2.3.4"
openai = "^1.35.14"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
typer = "^0.12.3"
pytest = "^8.2.2"
pyinstaller = "^6.9.0"

[tool.poetry.scripts]
aitk = "aitk.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"